live: Minimize media queue sizes.
At 24fps, each frame is ~41ms, so with just a couple in the queue
plus the frames that are in-flight on either end of the pipeline,
we're already looking at 100+ ms of latency.

Note that we would preferably have separate queues for audio and
video but that's a larger change.