live: More tolerant handling of decode errors.
Sometimes the decoder will return an error if a frame cannot be fully
processed, eg due to packet loss resulting in incomplete data that can't
be concealed.

In PyAV this throws an exception. If enough exceptions occur within a
short amount of time, the session is terminated.

However, the decoder can often continue processing the next frame. So
implement an error budget for this: for every error, increment
a counter up. If the counter hits a limit, raise an exception to
terminate the stream. If a frame was decoded without error, decrement
the error counter if it is nonzero.

If the stream is terminated due to the error budget being hit,
higher-level logic in media.py will also re-try the stream from the
beginning for good measure.