.github: Only tag images as latest for tags
We only want to release "latest" images* when
we are doing a release, not on every main build.

We have already changed staging to deploy from the
`-main` tags so it stays up to date with main, but
we only want prod to use a stable tag. Prod also
includes public Os, which should definitely only
deploy images that have been thoroughly tested and
tagged as a release.

* Keep in mind we're not referring to literally
`latest` docker tags, because we use the same docker
repository for multiple pipelines. So for example,
it's more like live-app-streamdiffusion being the
latest for the streamdiffusion pipelines.